Fix load-order issue that could cause overrides to error
