+++ server/routes/categories.js
@@ -5,8 +5,8 @@ const router = require('express-promise-router')()
 const log = require('../logger')
 const {getMeta} = require('../list')
 const {fetchDoc} = require('../docs')
-const {cleanName} = require('../utils')
 const {getTemplates, sortDocs, stringTemplate, formatUrl, pathPrefix} = require('../utils')
+const {cleanName} = require('../text')
 const {parseUrl} = require('../urlParser')
 
 router.get('*', handleCategory)
