+++ server/routes/playlists.js
@@ -5,7 +5,8 @@ const router = require('express-promise-router')()
 const log = require('../logger')
 const {getMeta, getPlaylist} = require('../list')
 const {fetchDoc} = require('../docs')
-const {stringTemplate, formatUrl, pathPrefix, cleanName} = require('../utils')
+const {stringTemplate, formatUrl, pathPrefix} = require('../utils')
+const {cleanName} = require('../text')
 const {parseUrl} = require('../urlParser')
 
 router.get('*', handlePlaylist)
